Use sledgehammer bindgen as Dioxus does
